refactor: Update member management by encapsulating functionality within Club model
